+++ examples/LibraryDemo.tsx
@@ -1,12 +1,14 @@
 import React, { useState } from "react";
+
 import ButtonVariantsDemo from "./ButtonDemo/ButtonVariantsDemo";
 import ProductCardDemo from "./ProductCardDemo/ProductCardDemo";
+import FilterPanelDemo from "./FilterPanelDemo/FilterPanelDemo";
 
 /**
  * Hlavní aplikace pro zobrazení všech příkladů
  */
 const LibraryDemo: React.FC = () => {
-  type ExampleType = "button" | "customization" | "productCard";
+  type ExampleType = "button" | "customization" | "productCard" | "filterPanel";
 
   const [activeExample, setActiveExample] = useState<ExampleType>("button");
 
@@ -16,6 +18,8 @@ const LibraryDemo: React.FC = () => {
         return <ButtonVariantsDemo />;
       case "productCard":
         return <ProductCardDemo />;
+      case "filterPanel":
+        return <FilterPanelDemo />;
       default:
         return <ButtonVariantsDemo />;
     }
@@ -56,6 +60,17 @@ const LibraryDemo: React.FC = () => {
           onClick={() => setActiveExample("productCard")}
         >
           Produkt
+        </button>
+        <button
+          className={`cursor-pointer py-2 px-4 rounded-md text-base transition-all duration-200 
+            ${
+              activeExample === "filterPanel"
+                ? "bg-green-800 text-white hover:bg-green-900"
+                : "bg-transparent text-gray-700 dark:text-gray-300 hover:bg-gray-100 dark:hover:bg-gray-800"
+            }`}
+          onClick={() => setActiveExample("filterPanel")}
+        >
+          Filtry
         </button>
       </nav>